Log format can be configured.
The format for each log can be configured with the configuration
option defined by $config['logs']['format']. %D is for
the time, %R is for the requester IP address, %H is for the host on
which the command has been executed and %C is for the command.

Warning! The configuration option to choose the logs file has been
moved from $config['misc']['logs'] to $config['logs']['file'].