Improve 'source-interface-id' option.
On software router, an IPv4 and IPv6 addresses need to be specified.
Not specifying one of them or both will result in the router trying
to use the best address to contact the destination.

This fix the bug where software routers could not ping or traceroute
IPv6 destination with only a IPv4 source address (obviously).