Fix traceroute for Cisco when using 'source'.
On Cisco, when using traceroute 'source' option with a FQDN, you
need to explicitly specify if you want an IPv4 or an IPv6
traceroute. To fix this we try to resolve the given FQDN and then
pass the IPv4 or IPv6 address instead of the FQDN

Also make sure that the FQDN is used in the issued command. Do not
use the resolved IP anymore.