Add new configuration options to enable/disable IPv6 or IPv4.
When disabling IPv6 or IPv4, the looking glass will not try to use both
IP protocol versions when looking for an AS routes or when pinging or
tracerouting.
If an IP address of a disabling IP version is used, the user will have an
error in return of its command. This error will tell him that the IP version
he is trying to use is disabled.

Please note that when trying to ping or traceroute a hostname from a Cisco or
a Juniper device, this looking glass will not prevent the device to use a
disabled IP version when executing its command.

This commit also contains some small changes with HTML tags.