Added Mikrotik ROS Traceroute fix
The Traceroute tool on Mikrotik ROS operates like an MTR and continually updates the console with metrics. This can be avoided with count=1 but if the traceroute encounters a blank hop for any reason, the router rewrites the console anyway. If running the traceroute with no controlling terminal, as in the case of an LG, the result gets output in chunks and only the final chunk is actually useful.

This update explodes the resultant string data into an array using the blank line between each chunk as a deliminator and then takes the final index as the resultant string.